# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Chavez API
  version: 0.1.0
  description: |
    API del harness Chavez: auth (Better Auth + magic link + device), vault de providers,
    workspaces/sessions/chats por HTTP, y sincronización vía WebSocket `/ws`.

    **Auth:** Bearer (`Authorization: Bearer <session-token>`) o cookie de sesión Better Auth.
    Sign-in y sign-up usan el mismo email (magic link).

    **WebSocket:** conectar a `ws://<host>/ws?token=<token>` y enviar mensajes JSON
    `{ type, id, ... }` → respuesta `{ type, id, ok, data?, error? }`.

    Las rutas bajo `/api/auth/*` las atiende Better Auth (catch-all). Los shapes documentados
    reflejan el uso en este proyecto; detalles pueden variar con la versión del plugin.
  contact:
    name: Chavez
servers:
  - url: http://localhost:25001
    description: Stack local (API fija :25001)
...
paths:
...
  /workspaces/:
    get:
      tags: [Workspaces]
      summary: Listar workspaces del usuario
      operationId: listWorkspaces
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Workspaces + openConnections
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceWithConnections"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workspaces/{workspaceId}/sessions:
    get:
      tags: [Workspaces, Sessions]
      summary: Sessions de un workspace
      operationId: listWorkspaceSessions
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace + sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentSession"
                  openConnections:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /connections:
    get:
      tags: [Connections]
      summary: Sockets WebSocket abiertos del usuario
      operationId: listConnections
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Lista de conexiones
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConnectionPublic"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Detalle de session
      operationId: getSession
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session + workspace resumen
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: "#/components/schemas/AgentSession"
                  workspace:
                    type: object
                    properties:
                      id:
                        type: string
                      path:
                        type: string
                      name:
                        type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{sessionId}/chats:
    get:
      tags: [Sessions, Chats]
      summary: Chats de una session
      operationId: listSessionChats
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Lista de chats
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  chats:
                    type: array
                    items:
                      $ref: "#/components/schemas/Chat"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /chats/{chatId}:
    get:
      tags: [Chats]
      summary: Chat + mensajes
      operationId: getChat
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ChatId"
      responses:
        "200":
          description: Chat y mensajes ordenados
          content:
            application/json:
              schema:
                type: object
                properties:
                  chat:
                    $ref: "#/components/schemas/Chat"
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Token de sesión Better Auth (CLI device token o session token)
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: |
        Cookie de sesión Better Auth (nombre exacto puede incluir prefijo/host).
        El hub web usa `credentials: include`.
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChatId:
      name: chatId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: Unauthorized
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        path:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceWithConnections:
      allOf:
        - $ref: "#/components/schemas/Workspace"
        - type: object
          properties:
            openConnections:
              type: integer
            daemonBound:
              type: boolean
            daemonHostname:
              type: string
              nullable: true
            daemonPath:
              type: string
    AgentSession:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Chat:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        chatId:
          type: string
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: |
            User attaches: `{ attachments: [{ path, kind, status, byteSize, mime,
            truncated, listing, hydratedText, sha256 }] }`. Never `imageBase64`.
            Tool rows: toolName, toolCallId, status, input, output.
        createdAt:
          type: string
          format: date-time
    ConnectionPublic:
      type: object
      properties:
        connectionId:
          type: string
        workspaceId:
          type: string
          nullable: true
        path:
          type: string
          nullable: true
        clientKind:
          type: string
          enum: [client, daemon]
        hostname:
          type: string
          nullable: true
          description: os.hostname() del proceso daemon; null en clientes web
        role:
          type: string
          enum: [primary, standby, client]
        connectedAt:
          type: string
          format: date-time

    # --- WebSocket protocol ---
...